sub asssy router query name refactored
