+++ src/lib/server.js
@@ -10,7 +10,7 @@ const loggerMiddleware = require('./logger-middleware');
 const errorMiddleware = require('./error-middleware');
 const subAssemblyRouter = require('../routes/subAssembly-routes');
 const partRouter = require('../routes/part-routes');
-const sendAssyRouter = require('../routes/sendAssyRoutes');
+const sendSubAssyRoutes = require('../routes/sendSubAssyRouter');
 
 const app = express();
 
@@ -21,7 +21,7 @@ app.use(cors({
 app.use(authRouter);
 app.use(accountRouter);
 app.use(subAssemblyRouter);
-app.use(sendAssyRouter);
+app.use(sendSubAssyRoutes);
 app.use(partRouter);
 
 // GLOBAL MIDDLEWARE
